Make module for showSection function
